Move nocode bootstrap code to more appropriate package name
