+++ instrumentation/nocode/bootstrap/src/main/java/io/opentelemetry/javaagent/bootstrap/nocode/NocodeEvaluation.java
@@ -3,11 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.bootstrap;
+package io.opentelemetry.javaagent.bootstrap.nocode;
 
 import java.util.concurrent.atomic.AtomicReference;
 
-// FIXME replace with inline cached expression
 public class NocodeEvaluation {
 
   public interface Evaluator {
